[feat]sAbstractApplicationDependencyRegistrar.cs拆分到多个分布类
