+++ src/ServerApi/Services/Shared/Adnc.Shared.Application/Registrar/AbstractApplicationDependencyRegistrar.Caching.cs
@@ -1,0 +1,39 @@
+﻿namespace Adnc.Shared.Application.Registrar;
+
+public abstract partial class AbstractApplicationDependencyRegistrar : IDependencyRegistrar
+{
+    /// <summary>
+    /// 注册Caching相关处理服务
+    /// </summary>
+    /// <param name="builder"></param>
+    protected virtual void AddCachingServices(Action<IServiceCollection> action = null)
+    {
+        action?.Invoke(Services);
+
+        Services.AddAdncInfraCaching(RedisSection);
+        var serviceType = typeof(ICachePreheatable);
+        var implTypes = ApplicationLayerAssembly.ExportedTypes.Where(type => type.IsAssignableTo(serviceType) && type.IsNotAbstractClass(true)).ToList();
+        if (implTypes.IsNotNullOrEmpty())
+        {
+            implTypes.ForEach(implType =>
+            {
+                Services.AddSingleton(implType, implType);
+                Services.AddSingleton(x => x.GetRequiredService(implType) as ICachePreheatable);
+            });
+        }
+    }
+
+    /// <summary>
+    /// 注册BloomFilter相关处理服务
+    /// </summary>
+    /// <param name="builder"></param>
+    protected virtual void AddBloomFilterServices(Action<IServiceCollection> action = null)
+    {
+        action?.Invoke(Services);
+
+        var serviceType = typeof(IBloomFilter);
+        var implTypes = ApplicationLayerAssembly.ExportedTypes.Where(type => type.IsAssignableTo(serviceType) && type.IsNotAbstractClass(true)).ToList();
+        if (implTypes.IsNotNullOrEmpty())
+            implTypes.ForEach(implType => Services.AddSingleton(serviceType, implType));
+    }
+}
